requirements.txt
atlassian-python-api>=3.41.16
beautifulsoup4>=4.12.3
httpx>=0.28.0
mcp>=1.0.0
python-dotenv>=1.0.1
markdownify>=0.11.6
